Separated to a library and a application
