+++ daq_system/src/rupdate/rupdate.cpp
@@ -37,6 +37,7 @@
 #define NDEBUG
 #include "debug.h"
 
+
 ClassImp(MyMainFrame)
 
 // ########################################################################
@@ -935,39 +936,4 @@ TH1* MyMainFrame::GetHistogramRow(int row)
 // ########################################################################
 // ########################################################################
 
-int main(int argc, char* argv[])
-{
-    TApplication theApp("ROOTupdate", &argc, argv);
 
-    gStyle->SetPadGridX(kTRUE);
-    gStyle->SetPadGridY(kTRUE);
-    gStyle->SetPalette(1);
-    gStyle->SetCanvasColor(10);
-    gStyle->SetOptStat(1111111);
-    gStyle->SetStatFormat("10.8g");
-    gStyle->SetNdivisions(520,"y");
-
-    // ------------------------------
-
-    int maxupdate = 2;
-
-    for( int opt=0; (opt = getopt(argc, argv, "u:")) != -1; ) {
-        switch (opt) {
-        case 'u':
-            maxupdate = atoi(optarg);
-            break;
-        default: /* '?' */
-            std::cerr << "Usage: " << argv[0] << "[-u update_secs]" << std::endl;
-            exit(EXIT_FAILURE);
-        }
-    }
-
-    // ------------------------------
-
-    spectra_attach_all(true);
-    new MyMainFrame(maxupdate);
-    theApp.Run(true);
-    spectra_detach_all();
-
-    return 0;
-}
